дополнение для круд

+++ go-http/internal/link/repository.go
@@ -2,6 +2,7 @@ package link
 
 import (
 	"app/adv-http/pkg/db"
+	"errors"
 )
 
 type LinkRepository struct {
@@ -40,9 +41,21 @@ func (repo *LinkRepository) GetByHash(hash string) (*Link, error) {
 }
 
 func (repo *LinkRepository) Delete(id uint) error {
-	result := repo.Database.DB.Delete(&Link{}, id)
-	if result.Error != nil {
-		return result.Error
+	finedLink, _ := repo.GetById(id)
+	if finedLink != nil {
+		result := repo.Database.DB.Delete(&Link{}, id)
+		if result.Error != nil {
+			return result.Error
+		}
 	}
-	return nil
+	return errors.New("link not found")
+}
+
+func (repo *LinkRepository) GetById(id uint) (*Link, error) {
+	var link Link
+	result := repo.Database.DB.First(&link, id)
+	if result.Error != nil {
+		return nil, result.Error
+	}
+	return &link, nil
 }
